requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
wavespeed>=0.1.0
requests>=2.31.0
psycopg2-binary>=2.9.0
boto3>=1.34.0
